+++ src/engraving/rendering/score/horizontalspacing.cpp
@@ -1341,7 +1341,9 @@ double HorizontalSpacing::minHorizontalDistance(const Segment* f, const Segment*
         // Multi-staff ChordBrackets are stored on their owner chords
         // and are therefore not fully represented in every visually spanned staff Shape.
         // Add the missing bracket collisions before combining the per-staff minimum distances.
-        ChordBracketLayout::updateHorizontalSpacing(f, ns, staffIdx, squeezeFactor, d);
+        if (f->isChordRestType() || ns->isChordRestType()) {
+            ChordBracketLayout::updateHorizontalSpacing(f, ns, staffIdx, squeezeFactor, d);
+        }
 
         if (f->isChordRestType() && ns->isChordRestType()) {
             checkCollisionsWithCrossStaffStems(f, ns, staffIdx, d);
@@ -1615,7 +1617,7 @@ void HorizontalSpacing::computeLyricsPadding(const Lyrics* lyrics1, const Engrav
 void HorizontalSpacing::computeChordBracketPadding(const EngravingItem* item1, const ChordBracket* chordBracket, double& padding)
 {
     const Chord* chord = chordBracket->chord();
-    const Chord* itemChord = static_cast<const Chord*>(item1->findAncestor(ElementType::CHORD));
+    const Chord* itemChord = toChord(item1->findAncestor(ElementType::CHORD));
     if (chord && itemChord && chord->segment() == itemChord->segment() && chord->part() == itemChord->part()) {
         // Padding a right-handed chord bracket to chords in the same Segment and part: use the same padding values as the left-handed case
         padding = item1->score()->paddingTable().at(ElementType::CHORD_BRACKET).at(item1->type());
